+++ package/Editor/Settings/DemystifySettingsProvider.cs
@@ -56,6 +56,10 @@ namespace Needle.Demystify
 				DrawActivateGUI(settings);
 				DrawSyntaxGUI(settings);
 
+				EditorGUILayout.Space(10);
+				EditorGUILayout.LabelField("Console Options", EditorStyles.boldLabel);
+				settings.Separator = EditorGUILayout.TextField(new GUIContent("Stacktrace Separator", "Adds a separator to Console stacktrace output between each stacktrace"), settings.Separator);
+
 				if(DemystifySettings.DevelopmentMode)
 				// using(new EditorGUI.DisabledScope(!settings.DevelopmentMode))
 				{