+++ package/Changelog.md
@@ -4,7 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.0.1-pre.3] - 2021-01-03
+## [1.0.1-pre.6] - 2021-03-03
+- fix/improve complex syntax highlighting for generic return types
+- prevent syntax highlighting being applied to message
+- fixed empty lines in Editor.log stacktraces
+
+## [1.0.1-pre.4] - 2021-01-03
+- improved theme color editing
 - updated EditorPatching dependency to 1.0.1
 - moved first install flag to project settings
 - try make debug log absolute paths relative